Restructure Component Folder in src
