+++ src/app/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App/index';
+import App from './components/App';
 // import '@babel/polyfill';
 
 ReactDOM.render(<App />, document.getElementById('root'));
